+++ CHANGELOG.md
@@ -6,6 +6,7 @@ FEATURES:
 
 * **New Resource:** `sonatypeiq_application_category` [GH-1]
 * **New Resource:** `sonatypeiq_config_crowd` [GH-1]
+* **New Resource:** `sonatypeiq_config_product_license` [GH-1]
 
 
 ## 0.11.0 July 30, 2025